CORS, XSS, IDOR

+++ backend/src/services/chat.service.js
@@ -1,8 +1,12 @@
 import prisma from "../config/prisma.js";
+import xss from "xss";
 
 export const createChat = async (userId, title) => {
   return prisma.chat.create({
-    data: { userId, title: title || "New Chat" },
+    data: { 
+      userId, 
+      title: xss(title) || "New Chat"
+    },
   });
 };
 
@@ -19,6 +23,21 @@ export const getChats = async (userId) => {
 };
 
 export const getChatById = async (chatId, userId) => {
+  const chat = await prisma.chat.findFirst({
+    where: {
+      id: chatId,
+    },
+  });
+
+  if (!chat) {
+    throw new Error("Chat not found");
+  }
+
+  // IDOR zaštita
+  if (chat.userId !== userId) {
+    throw new Error("Access denied");
+  }
+
   return prisma.chat.findFirst({
     where: {
       id: chatId,
@@ -36,12 +55,16 @@ export const deleteChat = async (chatId, userId) => {
   const chat = await prisma.chat.findFirst({
     where: {
       id: chatId,
-      userId: userId,
     },
   });
 
   if (!chat) {
-    throw new Error("Chat not found or access denied");
+    throw new Error("Chat not found");
+  }
+
+  // IDOR zaštita
+  if (chat.userId !== userId) {
+    throw new Error("Access denied");
   }
 
   return prisma.chat.delete({
